Unidad2/Practicas/requirements.txt
mysql-conector-python
